# src/main/resources/api/api-spec.yaml
openapi: 3.1.0
info:
  title: CRM Lead Management API
  version: 1.0.0
  description: REST API для управления лидами в CRM системе
servers:
  - url: http://localhost:8081
    description: Local development server

paths:
  /api/leads:
    get:
      operationId: getLeads
      summary: Получить список всех лидов
      tags:
        - Lead Management
      responses:
        '200':
          description: Успешное получение списка лидов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeadResponse'
    post:
      operationId: createLead
      summary: Создать нового лида
      tags:
        - Lead Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeadRequest'
      responses:
        '201':
          description: Лид успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadResponse'
        '400':
          description: Ошибка валидации

  /api/leads/{id}:
    get:
      operationId: getLeadById
      summary: Получить лида по ID
      tags:
        - Lead Management
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Лид найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadResponse'
        '404':
          description: Лид не найден
    put:
      operationId: updateLead
      summary: Обновить существующего лида
      tags:
        - Lead Management
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLeadRequest'
      responses:
        '200':
          description: Лид успешно обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadResponse'
        '404':
          description: Лид не найден
    delete:
      operationId: deleteLead
      summary: Удалить лида
      tags:
        - Lead Management
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Лид успешно удалён
        '404':
          description: Лид не найден

components:
  schemas:
    CreateLeadRequest:
      type: object
      required:
        - email
        - company
        - status
      properties:
        email:
          type: string
          format: email
          description: Email лида
        phone:
          type: string
          description: Телефон
        company:
          type: string
          description: Название компании
        status:
          type: string
          description: Статус лида

    UpdateLeadRequest:
      type: object
      required:
        - email
        - company
        - status
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
        company:
          type: string
        status:
          type: string

    LeadResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор лида
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        status:
          type: string